Factor out `moderation`
