+++ async-openai/src/types/mod.rs
@@ -11,6 +11,7 @@ mod image;
 mod message;
 mod message_file;
 mod model;
+mod moderation;
 mod run;
 mod step;
 mod thread;
@@ -25,6 +26,7 @@ pub use image::*;
 pub use message::*;
 pub use message_file::*;
 pub use model::*;
+pub use moderation::*;
 pub use run::*;
 pub use step::*;
 pub use thread::*;
